Finished Qualifiers, advancing to state!!!!
