+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AuraRobot.java
@@ -152,7 +152,7 @@ public class AuraRobot
     public static double HANG_NUM_MOTOR_ROTATIONS = 3.5;
 
 
-    public static double AUTO_WAIT_FOR_START = 5;
+    public static double AUTO_WAIT_FOR_START = 8;
     public static double AUTO_WAIT_FOR_OUTTAKE = 1.0;
     public static double AUTO_WAIT_FOR_YELLOW_DROP = 2.0;
     public static double AUTO_WAIT_RETURN_TO_INTAKE = 3.0;
